Ensures there is only one syncing happening at the same time and other minor tweaks.

+++ playlist_downloader.rb
@@ -15,46 +15,78 @@ class PlaylistDownloader
   def self.download(list, dir, lyrics_dir, cover, imported)
     require 'fileutils'
 
-    songs = File.open(list).read.lines.select { |x| !x.strip.empty? }.map { |x| x.strip }.uniq.reverse
-    imp = File.open(imported).read.lines.select { |x| !x.strip.empty? }.map { |x| x.strip }
+    trigger_file = "/tmp/xiami_sync.trigger"
 
-    FileUtils.rm_rf(dir)
-    FileUtils.mkdir_p(dir)
-    FileUtils.rm_rf(lyrics_dir)
-    FileUtils.mkdir_p(lyrics_dir)
+    loop do 
+      FileUtils.rm(trigger_file) if File.exists?(trigger_file)
 
-    songs.each do |s|
-      info = SongDownloader.download_info(s)
-      basename = "#{info.search('artist').text} - #{info.search('title').text}"
-      counter = 0
+      File.open("/tmp/xiami_sync.lock", 'w') do |lckf|
 
-      while File.exists?(File.join(dir, self.combine_name(basename, counter) + ".mp3")) do
-        counter += 1
+        unless lckf.flock(File::LOCK_NB | File::LOCK_EX)
+          puts "已有同步进程执行中。"
+          return false
+        end
+
+        puts "开始同步"
+
+        songs = File.open(list) do |f|
+          f.read.lines.select { |x| !x.strip.empty? }.map { |x| x.strip }.uniq.reverse
+        end
+
+        imp = File.open(imported) do |f|
+          f.read.lines.select { |x| !x.strip.empty? }.map { |x| x.strip }
+        end
+
+        FileUtils.rm_rf(dir)
+        FileUtils.mkdir_p(dir)
+        FileUtils.rm_rf(lyrics_dir)
+        FileUtils.mkdir_p(lyrics_dir)
+
+        songs.each do |s|
+          next unless s.to_i > 0
+          next if imported && imp.include?(s)
+
+          info = SongDownloader.download_info(s)
+          basename = "#{info.search('artist').text} - #{info.search('title').text}"
+          counter = 0
+
+          while File.exists?(File.join(dir, self.combine_name(basename, counter) + ".mp3")) do
+            counter += 1
+          end
+
+          filename = "#{self.combine_name(basename, counter)}.mp3"
+          lyrics_fn = "#{self.combine_name(basename, counter)}.lrc"
+          puts "正在下载 #{filename}"
+          SongDownloader.download(s, dir, info, filename)
+          lyrics = SongDownloader.retrieve_lyrics(s, info)
+
+          if !lyrics.strip.empty?
+            File.open(File.join(lyrics_dir, lyrics_fn), 'w') { |f| f.write(lyrics) }
+          end
+
+          path = File.join(dir, filename)
+
+          self.write_info(path, info, cover)
+
+          if imported && !imp.include?(s)
+            self.import_to_itunes(path)
+            imp += [s]
+            File.open(imported, 'w') { |f| f.write(imp.join("\n")) }
+            puts "已将 #{filename} 导入 iTunes"
+
+            yield :imported, {id: s, filename: filename, info: info} if block_given?
+          end
+        end
+
+        puts "同步完成"
+
       end
 
-      filename = "#{self.combine_name(basename, counter)}.mp3"
-      lyrics_fn = "#{self.combine_name(basename, counter)}.lrc"
-      puts "正在下载 #{filename}"
-      SongDownloader.download(s, dir, info, filename)
-      lyrics = SongDownloader.retrieve_lyrics(s, info)
-
-      if !lyrics.strip.empty?
-        File.open(File.join(lyrics_dir, lyrics_fn), 'w') { |f| f.write(lyrics) }
-      end
-
-      path = File.join(dir, filename)
-
-      self.write_info(path, info, cover)
-
-      if imported && !imp.include?(s)
-        self.import_to_itunes(path)
-        imp += [s]
-        File.open(imported, 'w').write(imp.join("\n"))
-        puts "已将 #{filename} 导入 iTunes"
-
-        yield :imported, {id: s, filename: filename, info: info}
-      end
+      break unless File.exists?(trigger_file)
     end
+
+    return true
+
   end
 
   def self.import_to_itunes(path)
